リファクタリング, stream ready を見てdetectionをするか判定
